% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_nx.R
\name{AUC_ln_K}
\alias{AUC_ln_K}
\title{AUC_ln_K}
\usage{
AUC_ln_K(R_NX)
}
\arguments{
\item{R_NX}{The R_NX curve, a vector of values}
}
\value{
A value, the area under the curve.
}
\description{
Area under the R_NX(K) curve when K is put on a logarithmic scale.
}
\author{
Guido Kraemer
}
