% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coranking-package.R
\docType{package}
\name{coRanking-package}
\alias{coRanking-package}
\title{coRanking: A package for calculating the co-ranking matrix
This package provides methods for the calculation of the
co-ranking matrix and derived measures to assess the quality of a
dimensionality reduciton}
\description{
coRanking: A package for calculating the co-ranking matrix
This package provides methods for the calculation of the
co-ranking matrix and derived measures to assess the quality of a
dimensionality reduciton
}
\references{
TODO: add the references
}

