% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{image.coranking}
\alias{image.coranking}
\title{Image function for the co-ranking matrix}
\usage{
\method{image}{coranking}(Q, ...)
}
\arguments{
\item{Q}{of class \code{coranking}.}

\item{...}{parameters for the image function.}

\item{legend}{if \code{T} plot a legend.}
}
\description{
Plots the co-ranking matrix nicely
}
\details{
Plots the co-ranking matrix nicely for visual inspection.  uses the
\code{image} function internaly, \code{...} is passed down to the image
function. The values in the co-ranking matrix are logscaled for
better contrast.
}
\author{
Guido Kramer
}

