% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_nx.R
\name{plot_R_NX}
\alias{plot_R_NX}
\title{Plot the R_NX curve}
\usage{
plot_R_NX(R_NXs, pal = grDevices::palette(), ylim = c(0, 0.9))
}
\arguments{
\item{R_NXs}{A list of R_NX curves, names from the list will appear in the
legend.}

\item{pal}{a vector of colors}

\item{ylim}{set the y-axis limits of the plot}
}
\value{
A ggplot object.
}
\description{
Produces a plot with the R_NX curves from the arguments
}
\author{
Guido Kraemer
}
