% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_nx.R
\name{Q_NX}
\alias{Q_NX}
\title{The \eqn{Q_{NX}(K)} criterion}
\usage{
Q_NX(Q)
}
\arguments{
\item{Q}{a co-ranking matrix}
}
\value{
A vector with the values for Q_NX(K)
}
\description{
A curve indicating the percentage of points that are mild in- and extrusions
or keep their rank.
}
\details{
\deqn{ Q_{NX}(K) = \frac{1}{KN} \sum_{k=1}^{K}\sum_{l=1}^{K}Q_{kl} }
}
\references{
Lueks, W., Mokbel, B., Biehl, M., & Hammer, B. (2011). How to
  Evaluate Dimensionality Reduction? - Improving the Co-ranking Matrix.
  ArXiv:1110.3917 [Cs]. http://arxiv.org/abs/1110.3917
}
\author{
Guido Kraemer
}
