% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_nx.R
\name{R_NX}
\alias{R_NX}
\title{The R_NX criterion}
\usage{
R_NX(Q)
}
\arguments{
\item{Q}{a co-ranking matrix}
}
\value{
A vector with the values for R_NX(K)
}
\description{
A curve indicating the improvement of the embedding over a random embedding.
Values range from 0, for a random embedding, to 1 for a perfect embedding.
}
\details{
\deqn{ Q_{NX}(K) = \sum_{1\leq k\leq K}\sum_{1\leq l\leq K} \frac{q_{kl}}{KN} }
Counts the upper left K-by-K block of Q, i.e. it considers the preserved
ranks on the diagonal and the permutations within a neighborhood.

\deqn{ R_{NX}(K) = \frac{(N-1)Q_{NX}(K)-K}{N-1-K} } A resulting vale of 0
corresponds to a random embedding, a value of 1 to a perfect embedding of the
K-ary neighborhood.
}
\author{
Guido Kraemer
}
