% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankmatrix.R
\name{rankmatrix}
\alias{rankmatrix}
\title{Rank matrix}
\usage{
rankmatrix(X, input = c("data", "dist"), use = "C")
}
\arguments{
\item{X}{data, dist object, or distance matrix}

\item{input}{type of input}

\item{use}{if 'C' uses the compiled library, else uses the native
R code}
}
\value{
returns a matrix of class \code{'rankmatrix'}
}
\description{
Replaces the elements of \eqn{X} with their rank in the column vector of
the distance matrix
}
\details{
Each column vector in the distance matrix (or the distance matrix
computed from the input) is replaced by a vector indicating the
rank of the distance inside that vector.

This is a computation step necessary for the co-ranking matrix and
provided mainly so that the user has the possibility to save
computation time.
}
\author{
Guido Kraemer
}

