% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_nx.R
\name{AUC_ln_K}
\alias{AUC_ln_K}
\title{AUC_ln_K}
\usage{
AUC_ln_K(R_NX)
}
\arguments{
\item{R_NX}{The R_NX curve, a vector of values}
}
\value{
A value, the area under the curve.
}
\description{
Area under the R_NX(K) curve when K is put on a logarithmic scale.
}
\references{
Lee, J. A., Peluffo-Ordóñez, D. H., & Verleysen, M., 2015. Multi-scale
  similarities in stochastic neighbour embedding: Reducing dimensionality
  while preserving both local and global structure. Neurocomputing, 169,
  246–261. https://doi.org/10.1016/j.neucom.2014.12.095
}
\author{
Guido Kraemer
}
