% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coranking.R
\name{coranking}
\alias{coranking}
\title{Co-Ranking Matrix}
\usage{
coranking(Xi, X, input = c("data", "dist", "rank"), use = "C")
}
\arguments{
\item{Xi}{high dimensional data}

\item{X}{low dimensional data}

\item{input}{type of input (see. details)}

\item{use}{\code{R} or \code{C} backend}
}
\value{
a matrix of class \code{'coranking'}
}
\description{
Calculate the co-ranking matrix to assess the quality of a
diminsionality reduction.
}
\details{
Calculate the coranking matrix, to assess the quality of a
dimensionality reduction.  \code{Xi} is input in high dimensions,
\code{X} is input in low dimensions the type of input is given in
\code{input}, if \code{input = 'data'} it will be transformed into
a distance matrix with the \code{dist} function, id \code{input ==
'rank'}, \code{Xi} and \code{X} are expected do be rank matrices.
}
\seealso{
\code{\link{rankmatrix}}
}
\author{
Guido Kraemer
}
