% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_nx.R
\name{plot_R_NX}
\alias{plot_R_NX}
\title{Plot the \eqn{R_{NX}(K)} curve}
\usage{
plot_R_NX(R_NXs, pal = grDevices::palette(), ylim = c(0, 0.9), ...)
}
\arguments{
\item{R_NXs}{A list of R_NX curves, names from the list will appear in the
legend}

\item{pal}{a vector of colors}

\item{ylim}{set the y-axis limits of the plot}

\item{...}{options for the plotting function}
}
\value{
Nothing, produces a plot.
}
\description{
Produces a plot with the \eqn{R_{NX}(K)} curves from the arguments
}
\references{
Lee, J. A., Peluffo-Ordóñez, D. H., & Verleysen, M., 2015. Multi-scale
  similarities in stochastic neighbour embedding: Reducing dimensionality
  while preserving both local and global structure. Neurocomputing, 169,
  246–261. https://doi.org/10.1016/j.neucom.2014.12.095
}
\author{
Guido Kraemer
}
